16bit redpitaya procedure paragraph moved up
the procedure to adapt the code for redpitaya 16 was located at the bottom of the tuto. Now it is located in the connecting blocks to the FPGA so that people will be aware that they can't use the same code for the RP14 and the RP16 before generating the bitstream.